Re-organize test classes for more readability
